Refactoring code for making a sepparate tools class
