+++ src/main/java/voicemod/websites/SupportPage.java
@@ -7,14 +7,14 @@ import static com.codeborne.selenide.Selenide.$;
 import static com.codeborne.selenide.Selenide.$x;
 
 public class SupportPage {
-    public String url = "https://support.voicemod.net/hc/es";
+    public String url = "https://support.voicemod.net/";
 
-    public SelenideElement contactSupportLink =  $x(
+    public SelenideElement contactSupportLink = $x(
             "//a[contains(@class, 'submit-a-request') " +
-            "and text() = 'Enviar una solicitud']");
+                    "and text() = 'Enviar una solicitud']");
     public SelenideElement label_email_text_box = $x(
             "//label[contains(@for, 'request_anonymous_requester_email') " +
-            "and text() = 'Correo electrónico']");
+                    "and text() = 'Correo electrónico']");
     public SelenideElement form_email_text_box = $(byId("request_anonymous_requester_email"));
     public SelenideElement label_request_subject_text_box = $x(
             "//label[contains(@for, 'request_subject') " +
